Adds Google's GIC driver, with changes to support AArch32.
This helped me find a bug in my IRQ handler, where Armv7-R and Armv8-R AArch32 both state that an IRQ exception is given an address in LR that is four bytes larger than the actual return address. Because of pipeline issues back in the mid-1980s or something.